==> 11_GMMAT_ART/1_Make_list_of_SNPs.ipynb <==
# Produce a list of SNPs that are found within 500bp of GMMAT ART candidates

library(data.table)
library(dplyr)

art_hits <- fread("/mnt/data/NSF_GWAS/notebooks/InPlantaGWAS/13_GMMAT_ART/chr_pos_from_7-OUT_QTLs_pass2_a2_in_vitro_included-1_modified_for_inplanta_paper.txt")

head(art_hits)

all_SNPs <- fread("../01_SNP_format_conversions//1323_cohort_maf01_geno10_contiguous.snp.pass.bim")

head(all_SNPs)

levels(factor(all_SNPs$V3))

colnames(all_SNPs) <- c("CHR", "SNP", "zero", "POS", "REF", "ALT")

df <- data.frame()

for(i in 1:nrow(art_hits)){
    this_CHR <- art_hits$CHR[i]
    this_POS <- art_hits$POS[i]
    
    these_hits <- all_SNPs[which(
    all_SNPs$CHR == this_CHR &
    all_SNPs$POS > (this_POS - 501) & 
    all_SNPs$POS < (this_POS + 501))]
    
    df <- bind_rows(df, these_hits)
}

nrow(df)

head(df)

outname <- "SNP_list_inplanta_regeneration_hits_from_ART_within_500bp.txt"

fwrite(as.data.table(df$SNP), outname, row.names = FALSE, col.names = FALSE)
