==> 11_GMMAT_ART/2_Extract_SNPs.ipynb <==
# Use PLINK to extract SNPs within 500bp of a window peak from ART candidates

plink \
--bfile /mnt/data/NSF_GWAS/notebooks/InPlantaGWAS/01_SNP_format_conversions/1323_cohort_maf01_geno10_contiguous.snp.pass \
--extract SNP_list_inplanta_regeneration_hits_from_ART_within_500bp.txt --allow-extra-chr --make-bed \
--out subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits
