==> 11_GMMAT_ART/3_Convert_SNPs_to_GDS.ipynb <==
# Convert SNP file from `PLINK` binary to GDS

library(SNPRelate)

library(SeqArray)

## Convert directly from `PLINK` binaries to `SeqArray` GDS

SeqArray::seqBED2GDS("subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits.bed",
                     "subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits.fam",
                     "subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits.bim",
                     "subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits_a3.gds")
