==> 11_GMMAT_ART/4_GMMAT_based_of_previous_a9_ART_hits.ipynb <==
# Run GMMAT Wald test over top SNPs for traits with maf > 0.01

<div class="alert alert-block alert-info"> This notebook is an adaptation of the earlier GMMAT deployment notebook, to be used for ART hits only.</div>


source("../08_GMMAT/R/GMMAT_utils.R")
source("../08_GMMAT/R/misc_utils.R")

library(GMMAT)
library(biclust)
library(tools)
library(data.table)
library(CMplot)
library(ggplot2)
library(ggrepel)

## Load our data

### Set paths for kinship matrix and SNP matrix, which will be same for everything*

kinship_path <- "../07_GEMMA/output/1323_cohort_maf01_geno10.cXX.txt"

gds_path <- "../11_GMMAT_ART/subset_maf01_geno10_contiguous_within_500bp_inplanta_regen_ART_hits_a3.gds"

ncores <- 22

setwd("../08_GMMAT/")

outdir <- "Results/batch5_maf01_geno10_ART/"

if(!dir.exists(outdir)){
    dir.create(outdir,
              recursive = TRUE)
}

List of heritable traits

h2_table <- fread("/mnt/data/NSF_GWAS/notebooks/InPlantaGWAS/09_Data_mining/sorted_h2_table_with_raw_names_added.csv")

heritable_traits <- h2_table$raw_trait_name[which(
h2_table$h2 > 0.10)]

length(heritable_traits)

## Timepoint 4 was not taken in Ph. 1 or Ph. 3

covariate_path <- "../07_GEMMA/covariates/Stem_regen_diam_and_phases_sansPh1Ph3Ph8.cov"

phenotype_paths <- list.files("../05_Parsing_phenodata/pheno_files/stem_regen/",
                              pattern = "4w_binarized\\.header\\.",
                              full.names = TRUE)

for(phenotype in phenotype_paths){
    suppressWarnings(GMMAT_workflow(phenotype_path = phenotype,
                   covariate_path = covariate_path,
                   kinship_path = kinship_path,
                   gds_path = gds_path,
                   ncores = ncores,
                   n_SNPs = 30000,
                   #pthresh = 1e-4,
                   outdir = outdir))
}

## PC traits only exist for genotypes studied in all timepoints

Same covariate file we just used for timepoint 4...

covariate_path <- "../07_GEMMA/covariates/Stem_regen_diam_and_phases_sansPh1Ph3Ph8.cov"

phenotype_paths <- list.files("../05_Parsing_phenodata/pheno_files/stem_regen",
                              pattern = "\\.binary\\.header",
                              full.names = TRUE)

desired_substrings <- heritable_traits

phenotype_paths_subset <- phenotype_paths[grep(paste(desired_substrings, collapse = "|"),
                                               phenotype_paths)]

phenotype_paths_subset <- phenotype_paths_subset[grepl("PC", phenotype_paths_subset)]

for(phenotype in phenotype_paths_subset[7:length(phenotype_paths_subset)]){
    suppressWarnings(GMMAT_workflow(phenotype_path = phenotype,
                   covariate_path = covariate_path,
                   kinship_path = kinship_path,
                   gds_path = gds_path,
                   ncores = ncores,
                   n_SNPs = 30000,
                   #pthresh = 1e-4,
                   outdir = outdir))
}

## Timepoint 5 was not taken in Ph. 1

covariate_path <- "../07_GEMMA/covariates/Stem_regen_diam_and_phases_sansPh1Ph8.cov"

phenotype_paths <- list.files("../05_Parsing_phenodata/pheno_files/stem_regen/",
                              pattern = "5w_binarized\\.header\\.",
                              full.names = TRUE)

for(phenotype in phenotype_paths[c(1,3)]){
    GMMAT_workflow(phenotype_path = phenotype,
                   covariate_path = covariate_path,
                   kinship_path = kinship_path,
                   gds_path = gds_path,
                   ncores = ncores,
                   n_SNPs = 30000,
                   #pthresh = 1e-4,
                   outdir = outdir)
}

## Timepoint 2 and 3 were taken for all phases.

However, note we still leave Ph. 8 out to avoid singular matrix. One phase must be assigned to intercept.

covariate_path <- "../07_GEMMA/covariates/Stem_regen_diam_and_phases_sansPh8.cov"

phenotype_paths <- c(list.files("../05_Parsing_phenodata/pheno_files/stem_regen/",
                                pattern = "2w_binarized\\.header\\.",
                                full.names = TRUE),
                     list.files("../05_Parsing_phenodata/pheno_files/stem_regen/",
                                pattern = "3w_binarized\\.header\\.",
                                full.names = TRUE))

for(phenotype in phenotype_paths){
    GMMAT_workflow(phenotype_path = phenotype,
                   covariate_path = covariate_path,
                   kinship_path = kinship_path,
                   gds_path = gds_path,
                   ncores = ncores,
                   n_SNPs = 30000,
                   #pthresh = 1e-4,
                   outdir = outdir)
}
